requirements.txt
langchain-google-genai 
langchain-google-vertexai 
langgraph 
python-dotenv 
langchain 
langsmith 
langfuse 
fastapi 
uvicorn 
python-multipart 
mistune 
sqlalchemy 
psycopg2-binary 
pgvector
scikit-learn
newspaper4k
beautifulsoup4
requests
lxml
google-cloud-storage